pyproject.toml


[project]
name = "agentic-fastapi"
version = "0.1.0"
description = "FastAPI application with agentic behavior using LangGraph"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "pydantic[email]>=2.7.4,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "sqlalchemy",
    "python-dotenv==1.0.0",
    "python-multipart",
    "langgraph",
    "langchain",
    "openai",
    "google-cloud-aiplatform",
    "pyjwt==2.8.0",
    "bcrypt>=5.0.0",
    "langchain-google-genai",
    "langchain-openai",
    "langchain_mcp_adapters"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.0",
    "isort==5.13.2",
]